refactor: eventHandler 단일화
