+++ backend/src/main/java/com/deliveranything/domain/delivery/service/DeliveryService.java
@@ -7,8 +7,7 @@ import com.deliveranything.domain.delivery.entity.Delivery;
 import com.deliveranything.domain.delivery.enums.DeliveryStatus;
 import com.deliveranything.domain.delivery.event.dto.DeliveryStatusEvent;
 import com.deliveranything.domain.delivery.event.dto.OrderStatusUpdateEvent;
-import com.deliveranything.domain.delivery.handler.DeliveryStatusEventHandler;
-import com.deliveranything.domain.delivery.handler.OrderDeliveryStatusEventHandler;
+import com.deliveranything.domain.delivery.handler.DeliveryEventHandler;
 import com.deliveranything.domain.delivery.repository.DeliveryRepository;
 import com.deliveranything.domain.order.entity.Order;
 import com.deliveranything.domain.user.entity.profile.RiderProfile;
@@ -27,8 +26,7 @@ public class DeliveryService {
 
   private final RiderProfileService riderProfileService;
   private final DeliveryRepository deliveryRepository;
-  private final DeliveryStatusEventHandler deliveryStatusEventHandler;
-  private final OrderDeliveryStatusEventHandler orderDeliveryStatusEventHandler;
+  private final DeliveryEventHandler deliveryEventHandler;
   private final RedisTemplate<String, Object> redisTemplate;
 
   public void updateRiderStatus(Long riderId, RiderToggleStatusRequestDto riderStatusRequestDto) {
@@ -59,7 +57,7 @@ public class DeliveryService {
         .nextStatus(next)
         .build();
 
-    deliveryStatusEventHandler.handleDeliveryStatus(event);
+    deliveryEventHandler.handleDeliveryStatus(event);
   }
 
   // 라이더 배달 수락/거절 처리 및 상태 이벤트 발행
@@ -70,7 +68,7 @@ public class DeliveryService {
     // 이벤트만 발행 - 실제 상태 변경은 구독자에서 처리
     OrderStatusUpdateEvent event = new OrderStatusUpdateEvent(
         decisionRequestDto.orderId(), currentActiveProfileId, status);
-    orderDeliveryStatusEventHandler.handleOrderDeliveryStatus(event);
+    deliveryEventHandler.handleOrderDeliveryStatus(event);
   }
 
   // Delivery 생성
